Integrate networking into main file
